+++ MatchEntities/Models/PlayerRoundStats.cs
@@ -38,7 +38,7 @@ namespace MatchEntities
         }
 
         public long MatchId { get; set; }
-        public short Round { get; set; }
+        public byte Round { get; set; }
         public long PlayerId { get; set; }
         public int PlayedEquipmentValue { get; set; }
         public int MoneyInitial { get; set; }
